serverLLM/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
distro==1.9.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.60
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
openai==0.28.1
packaging==23.2
pandas==2.2.1
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tabula-py==2.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.1
yarl==1.4.2
zipp==3.17.0
